requirements.txt
flask
openai-whisper
transformers
huggingface-hub
torch
sentencepiece
